Use size_t in Zhou2015

+++ Zhou2015pf.hpp
@@ -155,13 +155,13 @@ void Zhou2015Labels(
 
   GridCellZ_pq<elev_t> priorityQueue;
 
-  for(int x=0;x<dem.viewWidth();x++){
+  for(size_t x=0;x<dem.viewWidth();x++){
     const int height = dem.viewHeight()-1;
     priorityQueue.emplace(x,0,     dem(x,0     ));
     priorityQueue.emplace(x,height,dem(x,height));
   }
 
-  for(int y=1;y<dem.viewHeight()-1;y++){
+  for(size_t y=1;y<dem.viewHeight()-1;y++){
     const int width = dem.viewWidth()-1;
     priorityQueue.emplace(0,    y,dem(0,    y));
     priorityQueue.emplace(width,y,dem(width,y));
@@ -202,22 +202,22 @@ void Zhou2015Labels(
   //knowing whether the tile has been flipped toe snure that we connect the
   //correct edges.
   if( ((edge & GRID_TOP)    && !flipV) || ((edge & GRID_BOTTOM) && flipV) )
-    for(int x=0;x<labels.viewWidth();x++)
+    for(size_t x=0;x<labels.viewWidth();x++)
       WatershedsMeet(labels(x,0),(label_t)1,dem(x,0),dem(x,0),my_graph);
 
   if( ((edge & GRID_BOTTOM) && !flipV) || ((edge & GRID_TOP)    && flipV) ){
     int bottom_row = labels.viewHeight()-1;
-    for(int x=0;x<labels.viewWidth();x++)
+    for(size_t x=0;x<labels.viewWidth();x++)
       WatershedsMeet(labels(x,bottom_row),(label_t)1,dem(x,bottom_row),dem(x,bottom_row),my_graph);
   }
 
   if( ((edge & GRID_LEFT)  && !flipH) || ((edge & GRID_RIGHT) && flipH) )
-    for(int y=0;y<labels.viewHeight();y++)
+    for(size_t y=0;y<labels.viewHeight();y++)
       WatershedsMeet(labels(0,y),(label_t)1,dem(0,y),dem(0,y),my_graph);  
 
   if( ((edge & GRID_RIGHT) && !flipH) || ((edge & GRID_LEFT)  && flipH) ){
     int right_col = labels.viewWidth()-1;
-    for(int y=0;y<labels.viewHeight();y++)
+    for(size_t y=0;y<labels.viewHeight();y++)
       WatershedsMeet(labels(right_col,y),(label_t)1,dem(right_col,y),dem(right_col,y),my_graph);
   }
